gitlab_webhooks/requirements.txt
docker-py
gitpython
